requirements.txt
matplotlib>=3.7.1
numpy>=1.26.4
opencv-python>=4.9
Pillow>=10.2.0
Pillow>=9.5.0
pyboy>=1.6.11
pytesseract>=0.3.10
scikit_learn>=1.4.0
skimage>=0.0
torch>=2.2.0
torchviz>=0.0.2tqdm>=4.65.0
